Change type of x and y to int8_t

+++ app/src/util/Position.cpp
@@ -5,17 +5,17 @@
 
 namespace Util {
 
-    Position::Position(int x, int y) : Vector2D { x, y } {}
+    Position::Position(int8_t x, int8_t y) : Vector2D { x, y } {}
 
     bool Position::isNextTo(const Position& other) const {
         if (other == *this) {
             return false;
         }
 
-        const int minX = x - 1;
-        const int maxX = x + 1;
-        const int minY = y - 1;
-        const int maxY = y + 1;
+        const int8_t minX = x - 1;
+        const int8_t maxX = x + 1;
+        const int8_t minY = y - 1;
+        const int8_t maxY = y + 1;
 
         return other.x >= minX && other.x <= maxX && other.y >= minY && other.y <= maxY;
     }
